+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changelog follow https://keepachangelog.com/ format.
     *   Added: `epy.issubclass` (like `issubclass` but does not raises error for
         non-types)
     *   Added: `epy.groupby`, like `itertools.groupby` but returns a `dict`.
+    *   Fix: `epy.StrEnum` raises better error message if invalid input.
 *   Other:
     *   Added: Guide on
         [API design principles](https://github.com/google/etils/blob/main/docs/api-design.md).